requirements.txt
attrs==22.2.0
aws-cdk-lib==2.60.0
aws-cdk.asset-awscli-v1==2.2.49
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.38
cattrs==22.2.0
constructs==10.1.218
exceptiongroup==1.1.0
jsii==1.73.0
publication==0.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
